requirements.txt
action-msgs==1.2.2
action-tutorials-interfaces==0.20.6
action-tutorials-py==0.20.6
actionlib-msgs==4.9.0
actuator-msgs==0.0.1
addict==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.10.1
aiortc==1.14.0
aiosignal==1.4.0
aiosqlite==0.21.0
ament-clang-format==0.12.14
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.11.0
APScheduler==3.11.0
asttokens==3.0.0
async-timeout==5.0.1
asyncua==1.1.8
attrs==25.4.0
av==16.0.1
backoff==2.2.1
blinker==1.9.0
boto3==1.35.60
botocore==1.35.99
builtin-interfaces==1.2.2
cachetools==5.5.2
catkin-tools==0.9.5
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
coloredlogs==15.0.1
comm==0.2.3
composition-interfaces==1.2.2
ConfigArgParse==1.7.1
contourpy==1.3.2
control-msgs==4.8.0
controller-manager==2.52.0
controller-manager-msgs==2.52.0
cryptography==45.0.7
cv-bridge==3.2.1
cycler==0.12.1
Cython==3.0.12
dash==3.2.0
dataclasses-json==0.6.7
defusedxml==0.7.1
demo-nodes-py==0.20.6
diagnostic-msgs==4.9.0
diagnostic-updater==4.0.6
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.115.14
fastjsonschema==2.21.2
filelock==3.17.0
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
generate-parameter-library-py==0.5.0
geometry-msgs==4.9.0
google-crc32c==1.7.1
gps-msgs==2.1.1
graph-msgs==0.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
image-geometry==3.2.1
imageio==2.37.0
importlib_metadata==8.7.0
inference==0.59.0
iniconfig==2.3.0
interactive-markers==2.3.2
ipython==8.37.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
laser-geometry==2.4.1
launch==1.0.11
launch-param-builder==0.1.1
launch-pytest==1.0.11
launch-ros==0.19.12
launch-testing==1.0.11
launch-testing-ros==0.19.12
launch-xml==1.0.11
launch-yaml==1.0.11
lazy_loader==0.4
lifecycle-msgs==1.2.2
logging-demo==0.20.6
lxml==6.0.2
map-msgs==2.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
message-filters==4.3.11
moveit-configs-utils==2.5.9
moveit-msgs==2.2.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.8.0
nav-msgs==4.9.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.25.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
onnxruntime==1.21.1
onvif-zeep-async==2.0.0
openai==1.109.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
osrf-pycommon==2.1.6
packaging==24.2
paho-mqtt==1.6.1
pandas==2.3.3
parso==0.8.5
pcl-msgs==1.0.0
pendulum-msgs==0.20.6
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
polars==1.34.0
polars-runtime-32==1.34.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.2
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybase64==1.0.2
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.7.1
pydantic_core==2.33.2
pydot==2.0.0
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pylibsrtp==1.0.0
pylogix==1.0.5
pymodbus==3.8.3
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyquaternion==0.9.9
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-qt-binding==1.1.3
pytz==2025.2
PyYAML==6.0.3
qrcode==8.0
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.6
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.7
redis==5.0.8
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
resource-retriever==3.1.3
retrying==1.4.2
rich==13.9.4
rmw-dds-common==1.6.0
ros-gz-interfaces==0.244.20
ros2action==0.18.14
ros2bag==0.15.15
ros2cli==0.18.14
ros2component==0.18.14
ros2controlcli==2.52.0
ros2doctor==0.18.14
ros2interface==0.18.14
ros2launch==0.19.12
ros2lifecycle==0.18.14
ros2multicast==0.18.14
ros2node==0.18.14
ros2param==0.18.14
ros2pkg==0.18.14
ros2run==0.18.14
ros2service==0.18.14
ros2test==0.4.0
ros2topic==0.18.14
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.6
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.3
rosidl-typesupport-fastrtps-cpp==2.2.3
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpds-py==0.27.1
rpyutils==0.2.2
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.8
rqt-gui-py==1.1.8
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.8
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
s3transfer==0.10.4
scikit-image==0.25.2
scipy==1.15.3
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
shapely==2.0.7
shellingham==1.5.4
simple-pid==2.0.1
six==1.17.0
slack_sdk==3.33.5
sniffio==1.3.1
sortedcontainers==2.4.0
srdfdom==2.0.7
sros2==0.10.7
stack-data==0.6.3
starlette==0.46.2
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
structlog==24.4.0
supervision==0.26.1
sympy==1.14.0
teleop-twist-keyboard==2.4.1
tf2-geometry-msgs==0.25.16
tf2-kdl==0.25.16
tf2-msgs==0.25.16
tf2-py==0.25.16
tf2-ros-py==0.25.16
tf2-sensor-msgs==0.25.16
tf2-tools==0.25.16
theora-image-transport==2.5.4
threadpoolctl==3.6.0
tifffile==2025.5.10
tldextract==5.1.3
tokenizers==0.21.4
tomli==2.3.0
topic-monitor==0.20.6
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
traitlets==5.14.3
trajectory-msgs==4.9.0
triton==3.5.0
turtlesim==1.4.3
twilio==9.3.8
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.221
ultralytics-thop==2.0.17
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
urllib3==2.5.0
vision-msgs==4.1.1
visualization-msgs==4.9.0
wait_for2==0.3.2
wcwidth==0.2.14
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xacro==2.1.1
yarl==1.22.0
zeep==4.3.2
zipp==3.23.0
zxing-cpp==2.2.0
